코드 복구

+++ NGP_project/Client/BombObject.cpp
@@ -65,6 +65,6 @@ void BombObject::Move()
         break;
     }
 
-    _pos.x = std::clamp(_pos.x, float(gBackgroundRect.left), float(gBackgroundRect.right));
-    _pos.y = std::clamp(_pos.y, (float)(gBackgroundRect.top), (float)(gBackgroundRect.bottom));
+    _pos.x = std::clamp(_pos.x, (int)(gBackgroundRect.left), (int)(gBackgroundRect.right));
+    _pos.y = std::clamp(_pos.y, (int)(gBackgroundRect.top), (int)(gBackgroundRect.bottom));
 }
