+++ NGP_project/Client/Monster.cpp
@@ -36,19 +36,19 @@ void Monster::Init()
     switch (randSpawn(gen)) {
     case 0:
         // 북쪽
-        _pos = { float(gBackgroundRect.left + cellOffset), float(gBackgroundRect.top + CELL_SIZE / 2) };
+        _pos = { gBackgroundRect.left + cellOffset, gBackgroundRect.top + CELL_SIZE / 2 };
         break;
     case 1:
         // 동쪽
-        _pos = { float(gBackgroundRect.right - CELL_SIZE / 2), float(gBackgroundRect.top + cellOffset) };
+        _pos = { gBackgroundRect.right - CELL_SIZE / 2, gBackgroundRect.top + cellOffset };
         break;
     case 2:
         // 남쪽
-        _pos = { float(gBackgroundRect.left + cellOffset), float(gBackgroundRect.bottom - CELL_SIZE / 2) };
+        _pos = { gBackgroundRect.left + cellOffset, gBackgroundRect.bottom - CELL_SIZE / 2 };
         break;
     case 3:
         // 서쪽
-        _pos = { float(gBackgroundRect.left + CELL_SIZE / 2), float(gBackgroundRect.top + cellOffset) };
+        _pos = { gBackgroundRect.left + CELL_SIZE / 2, gBackgroundRect.top + cellOffset };
         break;
     default:
         break;
@@ -60,7 +60,7 @@ void Monster::Init()
 
 void Monster::FindTarget(GameObject* other)
 {
-    SetTargetPos({ float(randWidth(gen)), float(randHeight(gen)) });
+    SetTargetPos({ randWidth(gen), randHeight(gen) });
 }
 
 void Monster::Move()
